requirements.txt
pydantic
langchain
langchain-openai
langchain-anthropic
langchain-community
python-dotenv
wikipedia
duckduckgo-search